[glaze] update to 5.6.0

+++ ports/glaze/portfile.cmake
@@ -6,13 +6,14 @@ vcpkg_from_github(
     OUT_SOURCE_PATH SOURCE_PATH
     REPO stephenberry/glaze
     REF "v${VERSION}"
-    SHA512 23ddfebd05cb54a81677dc314a2b21c74cc2d35c362262d99178ecd64067aac2cb0cee14b1cd91de4d45c6980758f80e218d1d5f7e3f15f0bf00583b675ef8de
+    SHA512 8bacfaee61f98147dfea7cbb3c7a3feb074ea953a32e33f985ee21d83840e1937fbd2f92cd6306c8349b511839a400e111844e446d732d724307ed2b5b02348a
     HEAD_REF main
 )
 
 vcpkg_check_features(OUT_FEATURE_OPTIONS FEATURE_OPTIONS
     FEATURES
-        ssl glaze_ENABLE_SSL
+        interop glaze_BUILD_INTEROP
+        ssl     glaze_ENABLE_SSL
 )
 
 vcpkg_cmake_configure(
@@ -26,6 +27,8 @@ vcpkg_cmake_configure(
 vcpkg_cmake_install()
 vcpkg_cmake_config_fixup()
 
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug")
+if(NOT "interop" IN_LIST FEATURES)
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug")
+endif()
 
 vcpkg_install_copyright(FILE_LIST "${SOURCE_PATH}/LICENSE")
